New Development: No             JIRA Issue: No/Yes List JIRA ticket.       Ready to Release: Yes/No      Interface Changes: Yes/No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: Yes/No              Module(s): Module Names change impacts.            Description: yet more rm fun with python
git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@11188 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/xmlcasa/scripts/tests/test_plotms.py
@@ -19,8 +19,11 @@ class plotms_test(unittest.TestCase):
     plotfile_minsize = 120000
 
     def _cleanUp(self):
-        shutil.rmtree(self.ms)
-        shutil.rmtree(self.plotfile_jpg)
+        if os.path.exists(self.ms):
+            shutil.rmtree(self.ms)
+        if os.path.exists(self.plotfile_jpg):
+            os.remove(self.plotfile_jpg)
+
 
 
     def setUp(self):
